refactor(ToastNotification): move ToastNotification to folder
