+++ packages/components/src/Notification/ToastNotification/ToastNotification/index.ts
@@ -1,0 +1,1 @@
+export * from "./ToastNotification"
